+++ java-security-it/src/test/java/com/sap/cloud/security/test/integration/XsuaaIntegrationTest.java
@@ -5,6 +5,7 @@ import com.sap.cloud.security.config.Service;
 import com.sap.cloud.security.config.cf.VcapServicesParser;
 import com.sap.cloud.security.test.SecurityTestRule;
 import com.sap.cloud.security.token.Token;
+import com.sap.cloud.security.token.TokenClaims;
 import com.sap.cloud.security.token.validation.CombiningValidator;
 import com.sap.cloud.security.token.validation.ValidationResult;
 import com.sap.cloud.security.token.validation.validators.JwtValidatorBuilder;
@@ -50,6 +51,28 @@ public class XsuaaIntegrationTest {
 
 		ValidationResult result = tokenValidator.validate(token);
 		assertThat(result.isValid()).isTrue();
+	}
+
+	@Test
+	public void xsuaaTokenValidationFails_withIasCombiningValidator() throws IOException {
+		OAuth2ServiceConfiguration configuration = VcapServicesParser
+				.fromFile("/ias/vcapServices/serviceSingleBinding.json")
+				.withUrl("https://myauth.com")
+				.createConfiguration();
+
+		CombiningValidator<Token> tokenValidator = JwtValidatorBuilder.getInstance(configuration)
+				.build();
+
+		Token token = rule.getPreconfiguredJwtGenerator()
+				.fromFile("/xsuaa/tokens/userAccessTokenRSA256.json")
+				.withClaimValue(TokenClaims.XSUAA.CLIENT_ID, "T000310")
+				.withClaimValue(TokenClaims.ISSUER, "http://auth.com")
+				.createToken();
+
+		ValidationResult result = tokenValidator.validate(token);
+		assertThat(result.isValid()).isFalse();
+		assertThat(result.getErrorDescription()).startsWith(
+				"Issuer is not trusted because 'iss' 'http://auth.com' does not match host 'myauth.com' of the identity provider");
 	}
 
 }
